Make the LCS guard overflow-safe (round-6 review) (#2564)
GPT round-6 caught that the cell-cap guard cast to long AFTER the int addition
(`(long)(oldStream.Count + 1)`), so a Count of int.MaxValue overflows to negative
and bypasses the cap. Cast each operand to long before adding. The guard exists to
be robust against pathological input, so this closes the hole it was meant to plug.

Refs #2564.

+++ src/ILInspector.Evidence/EvidenceMatcher.cs
@@ -82,7 +82,7 @@ public static class EvidenceMatcher
         ArgumentNullException.ThrowIfNull(newStream);
         options ??= EvidenceMatchOptions.Default;
 
-        long cells = (long)(oldStream.Count + 1) * (newStream.Count + 1);
+        long cells = ((long)oldStream.Count + 1) * ((long)newStream.Count + 1);
         if (cells > MaxOrderedMatchCells)
         {
             throw new ArgumentException(
